Move utils to a new place
